Add simple genealogy structures
I'm thinking it may be a good idea to use BTreeMaps instead. Currently we construct a
network of nodes, but it's rather impractical. Maybe I can construct a network like so:

	children:
		name0 -> [name1, name2, name3, name4]
		name3 -> [name5, name6, name7]
	parents:
		name1 -> [name0]
		name2 -> [name0]
		name3 -> [name0]
		name4 -> [name0]
		name5 -> [name3]
		name6 -> [name3]
		name7 -> [name3]

Would this be easier to work with? It stores the same network, but without any tediousness. Also, numbers
can be used instead, this allows the use of multiplicities and easy renaming if we want to.

It also avoids the dreaded structure I have now. In addition, I think we can traverse the graph quite easily.
How does this work for finding common ancestors? We traverse through all parents (and stop after a certain
amount of follows). I think I'll change it.