Remove useless code
Currently, we can compute the coefficient of relationship, however, the coefficient of inbreeding is perhaps more interesting.

This is a recursive algorithm that needs to compute the coefficients of inbreeding. Luckily, memoization allows us to store the coefficient in the individuals.

Here's how the algorithm progresses:
1. Find all ancestors of A.
2. Find all non-crossing paths to its parents.
3. For these ancestors, compute their respective coefficient of inbreeding.
4. Use the coefficient and the paths to find the complete coefficient of inbreeding.

A missing case is where we have one ancestor that is wild, and the other part of an inbred tree.

What is then the coefficient of the descendant? In this case, there is no path between the two parents. I suppose: if the tree is far enough removed, that the coefficient approaches zero, so we'll use that.